backend/requirements.txt
langchain
langchain-community
langchain-experimental
langchain_groq
langchain_openai
langchain_google_community
langchain_tavily
langgraph
langgraph-checkpoint-sqlite
fastapi
python-dotenv
streamlit
uvicorn
pydantic
httpx
requests
langchain-google-community[places]